fix(chat): reconcile persisted stream completions
